refactor: refactor code based on code review feedback
